requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
biopandas==0.5.1
bioservices==1.12.1
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
cattrs==24.1.3
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
cpdb-protein==0.2.0
Cython==3.1.1
dash==3.0.4
debugpy==1.8.12
decorator==5.1.1
deepdiff==8.5.0
defusedxml==0.7.1
deprecation==2.1.0
easydev==0.13.3
et-xmlfile==1.1.0
fastjsonschema==2.21.1
Flask==3.0.3
fqdn==1.5.1
frozenlist==1.6.0
gevent==25.5.1
gitdb==4.0.12
GitPython==3.1.44
graphein==1.7.7
greenlet==3.2.2
grequests==0.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
importlib_metadata==8.6.1
ipykernel==6.29.5
ipyspeck==0.6.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isoduration==20.11.0
itsdangerous==2.2.0
jaxtyping==0.2.36
Jinja2==3.1.5
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_packaging==0.12.3
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
line_profiler==4.2.0
loguru==0.7.3
looseversion==1.1.2
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mistune==3.1.0
multidict==6.4.3
multipledispatch==1.0.0
narwhals==1.23.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
orderly-set==5.4.1
overrides==7.7.0
pandocfilters==1.5.1
parasail==1.3.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.1.1
prometheus_client==0.21.1
propcache==0.3.1
protobuf==5.29.3
psutil==6.1.1
MDAnalysis==2.7.0
numpy<2.0
py3Dmol==2.0.0.post2
pyarrow==19.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1

PyQt5==5.15.9
PyQt5-sip==12.12.2
python-dotenv==1.1.0
python-json-logger==3.2.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.1
requests-cache==1.2.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-click==1.8.9
rpds-py==0.22.3
Send2Trash==1.8.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stmol==0.0.9
streamlit==1.41.1
streamlit-molstar==0.4.21
suds-community==1.2.0
tenacity==9.0.0
terminado==0.18.1
tinycss2==1.4.0
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
url-normalize==2.2.1
watchdog==6.0.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
wget==3.2
widgetsnbextension==3.5.2
wrapt==1.17.2
xarray==2024.7.0
xmltodict==0.14.2
yarl==1.20.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
pytest==8.3.3
